+++ api/web-bff/src/main/java/com/zeroskill/web_bff/WebBffApplication.java
@@ -1,13 +1,13 @@
-package com.zeroskill.gateway;
+package com.zeroskill.web_bff;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class GatewayApplication {
+public class WebBffApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(GatewayApplication.class, args);
+        SpringApplication.run(WebBffApplication.class, args);
     }
 
 }
